requirements.txt
Cerberus==1.3.4
click==8.0.3
#dataclasses==0.8
Flask==2.0.2
Flask-Caching==1.10.1
Flask-Cors==3.0.10
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
psycopg2==2.9.1
PyJWT==2.3.0
six==1.16.0
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
flask-lambda-python36==0.1.0